Adding support for a new kind parameter
For our own exceptions we can set a 'type' using the 'kind' property.

If this is set on an exception we return it, otherwise its ignored (so plays well with any other exceptions).

For non-debug exceptions 'kind' may be changed to 'type' (I feel fine about this since we aren't offically returning an exception, just using an exception to automatically generate an error, which has no formal terms like exceptions do).