Moving per exception information into exceptions themselves
This is far more sustainable, less cluttered and stops this package becoming huge trying to keep tabs on every type of exception raised.

There will still be instances where it isn't practical to format an exception when it's thrown (within another package for example) but these are mich rarer than dealing with our own exceptions.

To enable additional to specified we now check for more (standardised) custom exception properties.